Library as a module in new file. CanvasWrapper now handles canvas and ctx
