+++ src/index.html
@@ -10,6 +10,6 @@
     <h1>MapMusic</h1>
     <canvas id="game" width="800" height="400"></canvas>
 
-    <script src="/out/main.js"></script>
+    <script type="module" src="/out/main.js"></script>
 </body>
 </html>
