+++ tsconfig.json
@@ -3,9 +3,12 @@
     "outDir": "./out",            // Pasta onde os arquivos compilados serão gerados
     "rootDir": "./src",           // (Opcional) Define a pasta onde estão os arquivos TypeScript
     "target": "esnext",           // Define a versão do JavaScript de saída
-    "module": "commonjs",         // Define o sistema de módulos
+    "module": "ESNext",           // Define o sistema de módulos
     "strict": true,               // Ativa verificações rigorosas
-    "sourceMap": true             // (Opcional) Cria mapas de código-fonte para facilitar o debug
+    "sourceMap": true,            // (Opcional) Cria mapas de código-fonte para facilitar o debug
+    "paths": {
+      "*": ["/node_modules/*", "/out/*", "/src/*"]
+    }
   },
   "include": ["src/**/*.ts"],     // Inclui todos os arquivos TypeScript da pasta src
   "exclude": ["node_modules"]     // Exclui a pasta node_modules
